+++ ModBuildConfig/ModVersionValidation.cs
@@ -1,0 +1,8 @@
+namespace Nickel.ModBuildConfig;
+
+internal enum ModVersionValidation
+{
+	Disabled,
+	Warning,
+	Error
+}